Crate xcursor

A crate to load cursor themes, and parse XCursor files.

Modules

parser

A module implementing XCursor file parsing.

Structs

CursorTheme

A cursor theme.

CursorThemeIml 🔒

SearchPathsEnvironment 🔒

Functions

expand_home_dir 🔒

If the first component of the path is ~, replaces it with the home dir. If no home dir is present, returns None.

parse_theme 🔒

Parse the content of the index.theme and return the Inherits value.

theme_inherits 🔒

Load the specified index.theme file, and returns a Some with the value of the Inherits key in it. Returns None if the file cannot be read for any reason, if the file cannot be parsed, or if the Inherits key is omitted.

theme_search_paths 🔒

Get the list of paths where the themes have to be searched, according to the XDG Icon Theme specification. If XCURSOR_PATH is set, it will override the default search paths.